+++ __tests__/server/ssrServer.spec.js
@@ -34,6 +34,7 @@ jest.mock('../../src/server/middleware/addFrameOptionsHeader');
 jest.mock('../../src/server/middleware/forwardedHeaderParser');
 jest.mock('../../src/server/utils/logging/serverMiddleware', () => (req, res, next) => setImmediate(next));
 jest.mock('../../src/universal/index');
+jest.mock('../../src/server/middleware/pwa', () => ({ serviceWorkerMiddleware: jest.fn(() => (req, res, next) => next()) }));
 jest.mock('../../mocks/scenarios', () => ({
   scenarios: true,
 }), { virtual: true });
@@ -97,6 +98,7 @@ describe('ssrServer', () => {
     let addCacheHeaders;
     let json;
     let forwardedHeaderParser;
+    let serviceWorker;
 
     function loadServer() {
       ({ json } = require('body-parser'));
@@ -116,6 +118,7 @@ describe('ssrServer', () => {
       addFrameOptionsHeader = require('../../src/server/middleware/addFrameOptionsHeader').default;
       addCacheHeaders = require('../../src/server/middleware/addCacheHeaders').default;
       forwardedHeaderParser = require('../../src/server/middleware/forwardedHeaderParser').default;
+      ({ serviceWorkerMiddleware: serviceWorker } = require('../../src/server/middleware/pwa'));
       const server = require('../../src/server/ssrServer').default;
 
       return server;
@@ -223,6 +226,15 @@ describe('ssrServer', () => {
         .get('/route')
         .end(() => {
           expect(addCacheHeaders).toBeCalled();
+          done();
+        });
+    });
+
+    it('should call service worker middleware', (done) => {
+      request(loadServer())
+        .get('/_/pwa/service-worker.js')
+        .end(() => {
+          expect(serviceWorker).toBeCalled();
           done();
         });
     });