+++ src/server/utils/onModuleLoad.js
@@ -23,6 +23,8 @@ import readJsonFile from './readJsonFile';
 import { extendRestrictedAttributesAllowList, validateSafeRequestRestrictedAttributes } from './safeRequest';
 import { setConfigureRequestLog } from './logging/serverMiddleware';
 import { setCreateSsrFetch } from './createSsrFetch';
+import { setEventLoopDelayThreshold } from './createCircuitBreaker';
+import { configurePWA } from '../middleware/pwa';
 
 // Trim build hash
 const { buildVersion } = readJsonFile('../../../.build-meta.json');
@@ -66,7 +68,7 @@ export default function onModuleLoad({
 }) {
   const {
     [CONFIGURATION_KEY]: {
-      // Tenant Root Specific
+      // Root Module Specific
       providedExternals,
       provideStateConfig,
       csp,
@@ -74,6 +76,8 @@ export default function onModuleLoad({
       configureRequestLog,
       extendSafeRequestRestrictedAttributes = {},
       createSsrFetch,
+      eventLoopDelayThreshold,
+      pwa,
       // Child Module Specific
       requiredExternals,
       validateStateConfig,
@@ -112,7 +116,9 @@ export default function onModuleLoad({
     extendRestrictedAttributesAllowList(extendSafeRequestRestrictedAttributes);
     setConfigureRequestLog(configureRequestLog);
     setCreateSsrFetch(createSsrFetch);
+    setEventLoopDelayThreshold(eventLoopDelayThreshold);
     logModuleLoad(moduleName, metaData.version);
+    configurePWA(pwa);
     return;
   }
 
@@ -130,9 +136,9 @@ export default function onModuleLoad({
       const providedExternal = RootModule[CONFIGURATION_KEY].providedExternals[externalName];
 
       if (!providedExternal) {
-        messages.push(`External '${externalName}' is required by ${moduleName}, but is not provided by tenant root module`);
+        messages.push(`External '${externalName}' is required by ${moduleName}, but is not provided by the root module`);
       } else if (!semver.satisfies(providedExternal.version, requestedExternalVersion)) {
-        messages.push(`${externalName}@${requestedExternalVersion} is required by ${moduleName}, but tenant root module provides ${providedExternal.version}`);
+        messages.push(`${externalName}@${requestedExternalVersion} is required by ${moduleName}, but the root module provides ${providedExternal.version}`);
       }
     });